Explicitly reset the on_stop stop-callback object before calling completion function in [exec.when.all] stae::complete effects.
It was considered too subtle to be relying on the destruction order of the members declared
in the exposition-only local `state` struct to guarantee that the stop-callback is destroyed before
the stop-source which the execution of the stop-callback might access.

It also wasn't clear whether the listed order of member declarations in the 'state' structure
was considered normative or whether we wanted to give implementations flexibility to declare
members in different orders.

The alternative considered was to add a note, or possibly something normative that requires
that the 'on_stop'  member is declared after the 'stop_src' member so that during destruction
of the operation-state the stop-callback is destroyed before the stop-source.